app/requirements.txt
pandas==2.3.1
chromadb==1.3.7
groq==0.37.1
langchain-groq==1.1.1
python-dotenv==1.1.1
semantic-router==0.1.12
streamlit==1.48.1